Fix the annotation processor
Our processor now generates code in `process` rather than in `finish`. This allows other processors
to process our generated code in the next round.

Fix #50
