+++ .gitignore
@@ -19,6 +19,7 @@ captures
 .idea/misc.xml
 .idea/androidTestResultsUserPreferences.xml
 .idea/appInsightsSettings.xml
+.idea/migrations.xml
 gradle.xml
 *.iml
